requirements.txt
# PPNR Risk Models - Python Dependencies
# Core scientific computing
numpy==1.24.3
pandas==2.0.3
scipy==1.11.1

# Statistical modeling and machine learning
scikit-learn==1.3.0
statsmodels==0.14.0
xgboost==1.7.6
lightgbm==4.0.0

# Time series analysis
arch==6.2.0
pmdarima==2.0.3

# Financial data and analysis
yfinance==0.2.18
fredapi==0.5.1
quantlib==1.31

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
dash==2.11.2

# Database and data handling
sqlalchemy==2.0.19
psycopg2-binary==2.9.7

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0
click==8.1.6

# Testing and development
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0

# Jupyter for analysis
jupyter==1.0.0
ipykernel==6.25.0

# PDF and report generation
weasyprint==59.0
reportlab==4.0.4
jinja2==3.1.2

# Additional data processing
openpyxl==3.1.2
xlsxwriter==3.1.2

# API clients
requests==2.31.0
urllib3==2.0.4

# Logging and monitoring
loguru==0.7.0

# Date and time utilities
python-dateutil==2.8.2